+++ config/initializers/rails_admin.rb
@@ -6,6 +6,9 @@ RailsAdmin.config do |config|
   require Rails.root.join('lib', 'rails_admin', 'changelog.rb')
   RailsAdmin::Config::Actions.register(RailsAdmin::Config::Actions::Changelog)
 
+  ## Custom Fields
+  require Rails.root.join('lib', 'rails_admin', 'config', 'fields', 'types', 'array_field.rb')
+  RailsAdmin::Config::Fields::Types::register(:array_field, RailsAdmin::Config::Fields::Types::ArrayField)
 
   ### Popular gems integration
 
@@ -75,7 +78,8 @@ RailsAdmin.config do |config|
   config.included_models = [
     'User',
     'PublicationLanguage','PublicationLanguage::Translation',
-    'Illustrator', 'Illustrator::Translation',
+    # 'Illustrator', 'Illustrator::Translation',
+    'Person', 'Person::Translation',
     'News', 'News::Translation',
     'Research', 'Research::Translation',
     'Tag', 'Tag::Translation',
@@ -98,6 +102,32 @@ RailsAdmin.config do |config|
   end
 
   config.model 'Illustrator::Translation' do
+    visible false
+    configure :locale, :hidden do
+      help ''
+    end
+    configure :is_public do
+      html_attributes required: required? && !value.present?, class: 'is-public-field'
+    end
+
+    include_fields :locale, :is_public, :name, :bio
+
+    edit do
+      field :bio, :ck_editor
+      fields :name, :bio do
+        help I18n.t('admin.help.required_for_publication')
+      end
+      # uploader's cannot make anything public
+      field :is_public do
+        visible do
+          !bindings[:view]._current_user.uploader?
+        end
+      end
+    end
+
+  end
+
+  config.model 'Person::Translation' do
     visible false
     configure :locale, :hidden do
       help ''